added exception handling

+++ src/main/java/com/mahreen/book_store/service/impl/BookServiceImpl.java
@@ -2,6 +2,7 @@ package com.mahreen.book_store.service.impl;
 
 import com.mahreen.book_store.dto.BookDto;
 import com.mahreen.book_store.entity.Book;
+import com.mahreen.book_store.exception.BookNotFoundException;
 import com.mahreen.book_store.mapper.BookMapper;
 import com.mahreen.book_store.repository.BookRepository;
 import com.mahreen.book_store.service.BookService;
@@ -9,6 +10,7 @@ import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class BookServiceImpl implements BookService {
@@ -20,16 +22,30 @@ public class BookServiceImpl implements BookService {
 
     @Override
     public BookDto getBook(String bookId) {
-        Book book = bookRepository.findBookById(bookId);
-        return BookMapper.toDto(book);
+        //Optional<T> is a container object introduced in Java 8.
+        //It’s used to represent a value that may or may not be present — instead of returning null.
+        Optional<Book> optionalBook = Optional.ofNullable(bookRepository.findBookById(bookId));
+
+        if (optionalBook.isPresent()) {
+            Book book = optionalBook.get();
+            return BookMapper.toDto(book);
+        }
+
+        else {
+            throw new BookNotFoundException("Book not found with ID: " + bookId);
+        }
     }
 
     @Override
     public List<BookDto> getAllBooks() {
         List<Book> books = bookRepository.findAll();
-        List<BookDto> bookDtoList = new ArrayList<>();
 
-        for (Book book : books){
+        if (books.isEmpty()) {
+            throw new BookNotFoundException("No books found.");
+        }
+
+        List<BookDto> bookDtoList = new ArrayList<>();
+        for (Book book : books) {
             bookDtoList.add(BookMapper.toDto(book));
         }
         return bookDtoList;
@@ -43,13 +59,28 @@ public class BookServiceImpl implements BookService {
 
     @Override
     public BookDto updateBookName(BookDto bookDto) {
-        bookRepository.updateBookNameByBookId(bookDto.bookId(), bookDto.name());
-        return this.getBook(bookDto.bookId());
+        Optional<Book> book = Optional.ofNullable(bookRepository.findBookById(bookDto.bookId()));
+
+        if(book.isPresent()) {
+            bookRepository.updateBookNameByBookId(bookDto.bookId(), bookDto.name());
+            return this.getBook(bookDto.bookId());
+        }
+        else {
+            throw new BookNotFoundException("Book not found with ID: " + bookDto.bookId());
+        }
     }
 
     @Override
     public void deleteBookByBookId(String bookId) {
 
-        bookRepository.deleteBookByBookId(bookId);
+        Optional<Book> book = Optional.ofNullable(bookRepository.findBookById(bookId));
+        if(book.isPresent()) {
+            bookRepository.deleteBookByBookId(bookId);
+        }
+
+        else {
+            throw new BookNotFoundException("Book not found with ID: " + bookId);
+        }
+
     }
 }
